+++ formguard/__init__.py
@@ -1,3 +1,3 @@
 """Invisible, layered form protection for Django."""
 
-__version__ = '0.8.0'
+__version__ = '0.8.1'