+++ formguard/contrib/turnstile/widgets.py
@@ -6,11 +6,12 @@ class TurnstileWidget(forms.Widget):
 
     template_name = 'formguard/contrib/turnstile/widget.html'
 
-    def __init__(self, site_key, theme='auto', size='normal', callback=None, attrs=None):
+    def __init__(self, site_key, theme='auto', size='normal', appearance=None, callback=None, attrs=None):
         super().__init__(attrs=attrs)
         self.site_key = site_key
         self.theme = theme
         self.size = size
+        self.appearance = appearance
         self.callback = callback
 
     def get_context(self, name, value, attrs):
@@ -18,6 +19,7 @@ class TurnstileWidget(forms.Widget):
         context['widget']['site_key'] = self.site_key
         context['widget']['theme'] = self.theme
         context['widget']['size'] = self.size
+        context['widget']['appearance'] = self.appearance
         context['widget']['callback'] = self.callback
         return context